Redo message pipe handling
Previously messages were sent in a pair - the first one was the prefix (e.g. `timed_cheat`) and the second one was the actual content (e.g. `invisible_cars:60000:Invisible Cars`).

Due to this the program was EXPECTING the content after the prefix message. If it couldn't find it, it could've run into issues.

This has been reworked to prevent race conditions (that I think were actually happening...)