Implement autosaving
Currently saving to Slot 8, like the CLEO script.
It has some issues when playing and loading a save with a higher amount of missions passed.
To compensate for this, the player can't click on "Start New Game" from the menu anymore if it was clicked once already (e.g. after starting the game)