Fix parachute bug when clearing weapons
When CJ is currently in the parachuting animation and the weapon is taken away, he's still gonna play that animation and when landing on the ground, will die but still be in the animation.

Before clearing weapons we now check if the player is in said animation, then we clear that task and remove all weapons.